Renamed application ID to com.android.mdl.app
