+++ appholder/src/main/java/com/android/mdl/app/viewmodel/TransferDocumentViewModel.kt
@@ -1,16 +1,16 @@
-package com.ul.ims.gmdl.appholder.viewmodel
+package com.android.mdl.app.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.security.identity.IdentityCredentialPresentation
 import androidx.security.identity.InvalidRequestMessageException
-import com.ul.ims.gmdl.appholder.transfer.TransferManager
-import com.ul.ims.gmdl.appholder.util.DocumentData.AAMVA_NAMESPACE
-import com.ul.ims.gmdl.appholder.util.DocumentData.DUMMY_CREDENTIAL_NAME
-import com.ul.ims.gmdl.appholder.util.DocumentData.MDL_DOCTYPE
-import com.ul.ims.gmdl.appholder.util.DocumentData.MDL_NAMESPACE
-import com.ul.ims.gmdl.appholder.util.TransferStatus
+import com.android.mdl.app.transfer.TransferManager
+import com.android.mdl.app.util.DocumentData.AAMVA_NAMESPACE
+import com.android.mdl.app.util.DocumentData.DUMMY_CREDENTIAL_NAME
+import com.android.mdl.app.util.DocumentData.MDL_DOCTYPE
+import com.android.mdl.app.util.DocumentData.MDL_NAMESPACE
+import com.android.mdl.app.util.TransferStatus
 
 class TransferDocumentViewModel(val app: Application) : AndroidViewModel(app) {
