+++ appholder/src/main/java/com/android/mdl/app/viewmodel/UserConsentViewModel.kt
@@ -1,10 +1,10 @@
-package com.ul.ims.gmdl.appholder.viewmodel
+package com.android.mdl.app.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
-import com.ul.ims.gmdl.appholder.transfer.TransferManager
-import com.ul.ims.gmdl.appholder.util.TransferStatus
+import com.android.mdl.app.transfer.TransferManager
+import com.android.mdl.app.util.TransferStatus
 
 class UserConsentViewModel(val app: Application) : AndroidViewModel(app) {
