+++ appholder/src/main/res/layout/fragment_user_consent.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="fragment"
-            type="com.ul.ims.gmdl.appholder.fragment.UserConsentFragment" />
+            type="com.android.mdl.app.fragment.UserConsentFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
