+++ appholder/src/main/res/layout/list_item_document.xml
@@ -26,7 +26,7 @@
 
         <variable
             name="document"
-            type="com.ul.ims.gmdl.appholder.document.Document" />
+            type="com.android.mdl.app.document.Document" />
     </data>
 
     <androidx.cardview.widget.CardView
